Fix for Nuttx and Posix building.

+++ src/modules/msp_osd/msp_osd.cpp
@@ -418,7 +418,7 @@ int MspOsd::print_status()
 	PX4_INFO("Running on %s", _device);
 	PX4_INFO("\tinitialized: %d", _is_initialized);
 	PX4_INFO("\tinitialization issues: %d", _performance_data.initialization_problems);
-	PX4_INFO("\tscroll rate: %ld", _param_scroll_rate.get());
+	PX4_INFO("\tscroll rate: %d", static_cast<int>(_param_scroll_rate.get()));
 	PX4_INFO("\tsuccessful sends: %lu", _performance_data.successful_sends);
 	PX4_INFO("\tunsuccessful sends: %lu", _performance_data.unsuccessful_sends);
